optimizations made:
Replaced the hash command with command -v, which is more portable and POSIX compliant.
Combined the apt-get update and apt-get install commands into a single command to reduce the number of commands executed.
Simplified the command to retrieve the system's IP address by using the ip -4 addr show command and filtering the output with grep and awk.
Removed the redundant grep command in the condition for checking the number of IPv4 addresses on the system.
Simplified the command for selecting the user's desired IPv4 address by using sed instead of grep and cut.